requirements.txt

folium==0.15.1
inflection==0.5.1
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
Pillow==10.4.0
streamlit==1.37.1
streamlit_folium==0.18.0
